Move Nation-related GameFile functions to new file OGFILE_NATION.cpp.
